% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_known_conflicts}
\alias{check_known_conflicts}
\title{Check for Known Package Conflicts}
\usage{
check_known_conflicts(status)
}
\arguments{
\item{status}{Current status list}
}
\value{
Updated status list
}
\description{
Check for Known Package Conflicts
}
\keyword{internal}
