% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_data_files}
\alias{copy_data_files}
\title{Copy data files to the Psych-DS data directory}
\usage{
copy_data_files(files, dest_dir)
}
\arguments{
\item{files}{Vector of file paths to copy}

\item{dest_dir}{Destination directory}
}
\value{
Character vector of copied file paths
}
\description{
Copy data files to the Psych-DS data directory
}
\keyword{internal}
