% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_datapackage}
\alias{create_datapackage}
\title{Create a datapackage.json file}
\usage{
create_datapackage(dir_path, files, dataset_info)
}
\arguments{
\item{dir_path}{Path to the directory where the file will be created}

\item{files}{List of data files}

\item{dataset_info}{List containing dataset information}
}
\value{
Invisible NULL
}
\description{
Create a datapackage.json file
}
\keyword{internal}
