% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dataset_description}
\alias{create_dataset_description}
\title{Create a dataset_description.json file}
\usage{
create_dataset_description(dir_path, dataset_info)
}
\arguments{
\item{dir_path}{Path to the directory where the file will be created}

\item{dataset_info}{List containing dataset information}
}
\value{
Invisible NULL
}
\description{
Create a dataset_description.json file
}
\keyword{internal}
