% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_psych_ds_structure}
\alias{create_psych_ds_structure}
\title{Create a Psych-DS directory structure}
\usage{
create_psych_ds_structure(base_dir, opt_dirs = NULL)
}
\arguments{
\item{base_dir}{Path to the base directory}

\item{opt_dirs}{Optional directories to create}
}
\value{
Invisible NULL
}
\description{
Create a Psych-DS directory structure
}
\keyword{internal}
