% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_data_dictionary}
\alias{generate_data_dictionary}
\title{Generate a data dictionary from CSV files}
\usage{
generate_data_dictionary(files)
}
\arguments{
\item{files}{Vector of CSV file paths}
}
\value{
List of data frames, one for each file
}
\description{
Generate a data dictionary from CSV files
}
\keyword{internal}
