% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_psych_ds_dir}
\alias{is_valid_psych_ds_dir}
\title{Check if a directory is a valid Psych-DS dataset}
\usage{
is_valid_psych_ds_dir(dir_path)
}
\arguments{
\item{dir_path}{Path to the directory to check}
}
\value{
Logical indicating if the directory is a valid Psych-DS dataset
}
\description{
Check if a directory is a valid Psych-DS dataset
}
\keyword{internal}
