% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{offer_installation_help}
\alias{offer_installation_help}
\title{Offer Installation Help}
\usage{
offer_installation_help(status, min_versions)
}
\arguments{
\item{status}{Status list}

\item{min_versions}{Minimum version requirements}
}
\description{
Offer Installation Help
}
\keyword{internal}
