% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{report_dependency_status}
\alias{report_dependency_status}
\title{Report Dependency Check Results}
\usage{
report_dependency_status(status, mode)
}
\arguments{
\item{status}{Status list from check_dependencies}

\item{mode}{Startup mode}
}
\description{
Report Dependency Check Results
}
\keyword{internal}
