% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_app_safe}
\alias{run_app_safe}
\title{Initialize Shiny App with Robust Package Management}
\usage{
run_app_safe(app_dir = system.file("app", package = "psychds"), ...)
}
\arguments{
\item{app_dir}{Directory containing the Shiny app}

\item{...}{Additional arguments passed to shiny::runApp}
}
\description{
Main entry point that ensures all dependencies are met
}
