% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_dataset}
\alias{validate_dataset}
\title{Validate a Psych-DS dataset}
\usage{
validate_dataset(dir_path)
}
\arguments{
\item{dir_path}{Path to the Psych-DS dataset directory}
}
\value{
List containing validation results
}
\description{
Validate a Psych-DS dataset
}
\keyword{internal}
